+++ pyout/elements.py
@@ -38,15 +38,31 @@ schema = {
             "default": False,
             "scope": "field"},
         "width": {
-            "description": "Width of field",
+            "description": """Width of field.  With the default value, 'auto',
+            the column width is automatically adjusted to fit the content and
+            may be truncated to ensure that the entire row fits within the
+            available output width.  An integer value forces all fields in a
+            column to have a width of the specified value. In addition, an
+            object can be specified.  Its 'min' and 'max' keys specify the
+            minimum and maximum widths allowed, whereas the 'width' key
+            specifies a fixed width.  The 'marker' key specifies the marker
+            used for truncation ('...' by default).  Where the field is
+            truncated can be configured with 'truncate': 'right' (default),
+            'left', or 'center.""",
             "oneOf": [{"type": "integer"},
                       {"type": "string",
                        "enum": ["auto"]},
                       {"type": "object",
                        "properties": {
-                           "auto": {"type": "boolean"},
                            "max": {"type": ["integer", "null"]},
-                           "min": {"type": ["integer", "null"]}}}],
+                           "min": {"type": ["integer", "null"]},
+                           "width": {"type": "integer"},
+                           "marker": {"type": ["string", "boolean"]},
+                           "truncate": {"type": "string",
+                                           "enum": ["left",
+                                                    "right",
+                                                    "center"]}},
+                       "additionalProperties": False}],
             "default": "auto",
             "scope": "column"},
         # Other style elements