+++ pyout/tests/test_field.py
@@ -8,7 +8,6 @@ import six
 from pyout.field import Field
 from pyout.field import Nothing
 from pyout.field import StyleProcessors
-from pyout.field import Truncater
 
 
 def test_field_base():
@@ -56,35 +55,6 @@ def test_something_about_nothing(text):
     assert "{:5}".format(nada) == "{:5}".format(text)
     assert "x" + nada == "x" + text
     assert nada + "x" == text + "x"
-
-
-def test_truncate_mark_true():
-    fn = Truncater(7, marker=True).truncate
-
-    assert fn(None, "abc") == "abc"
-    assert fn(None, "abcdefg") == "abcdefg"
-    assert fn(None, "abcdefgh") == "abcd..."
-
-
-def test_truncate_mark_string():
-    fn = Truncater(7, marker="…").truncate
-
-    assert fn(None, "abc") == "abc"
-    assert fn(None, "abcdefg") == "abcdefg"
-    assert fn(None, "abcdefgh") == "abcdef…"
-
-
-def test_truncate_mark_short():
-    fn = Truncater(2, marker=True).truncate
-    assert fn(None, "abc") == ".."
-
-
-def test_truncate_nomark():
-    fn = Truncater(7, marker=False).truncate
-
-    assert fn(None, "abc") == "abc"
-    assert fn(None, "abcdefg") == "abcdefg"
-    assert fn(None, "abcdefgh") == "abcdefg"
 
 
 def test_style_value_type():